Power of Hybrid: Edge computing and cloud using AI
Power of Hybrid: Edge computing and cloud using AI
The fourth industrial revolution has become the buzzword for the decade and is also here to stay.
But what makes this revolution so popular?
The fourth industrial revolution or Industry 4.0 delivers the promise of better production at lower costs, higher volumes, and improved quality.¬¬
But how does it deliver this promise?
Digital transformation can ease off the manufacturing processes for most of the factories and make them lean in nature. By integrating big data, having OT-IT connectivity, and implementing Internet of Things, smart manufacturing can be easily adopted by the manufacturers which can make them leverage Industry 4.0.
But how can the strategy be effective?
Often most of the digital manufacturers develop solutions in isolation by deploying AI in bits and pieces, in the absence of a strong organizational strategy. But to make digital transformation a reality, the most effective strategy involves combining edge computing with the hybrid clouds. This hybrid strategy can aid in scaling the value of artificial intelligence throughout the entire manufacturing business. For instance, the manufacturing connectivity and intelligence platforms in a company can be deployed on the assembly lines at the edge of the network in difference countries of the company. This system can be managed using the edge computing technology and hybrid clouds.
Additionally, this system provides an instant payback in comparison to the manual inspection of the manufacturing. This is because using AI, the inspection time can be reduced easily by a minimum of 10 minutes, for each use case. This also eases off the tasks for the different team mates as the nature of the task seems to be tedious to them.
The edge platforms and AI models can be controlled and managed from a central point through automated process and the cloud. This way the cost of maintenance for the software can simply go down by a striking 20%. This decline in the operating cost can save a huge amount of the capital expenditure on an annual basis. This simply highlights that edge computing can act as a catalyst for digital transformation by optimally using AI, big data and IoT.
There’s more to do with Industry 4.0
In the current times, the endpoint devices have become smarter than what most organizations know them for. The purpose of the sensors extends beyond sharing only vital data and information, as they come attached with their own memory resources and compute. These sensors are holistic computer systems that require regular software updates, constant management, along with security patches to keep them safe. It can be stated that using AI in sensors and other endpoints, real transformation for the manufacturing sector can happen at the network edge.
What is the role of edge computing?
Since some of the workloads are compute intensive, depending on huge amount of big data images and data, the AI models are deployed on the edge devices. For this purpose, the processing occurs at the edge in real time. As the data is collected at the edge, it is easier and faster to calculate the data at the edge itself. This way anomalies can be detected quickly and acted upon immediately, eliminating the scope of delays due to data transfer to the cloud. Therefor, Edge computing can simply remove the latency and bandwidth issues.
Even so, small little changes may be required on the manufacturing line only to improve the quality of the output. This may require newer AI models to deploy the changes onto the edge devices and retain the changes. This change implementation can be easy when done manually, but only when it is in a limited number. But if the change needs to be made at uncountable stations and edge devices, it may not be possible manually. For this reason, automation from edge computing and hybrid cloud comes handy, accelerating the change management.
Why Implement Brabo?
Brabo, an Industry 4.0 platform developed by Solulever, a Dutch technology startup provides top industrial connectivity. It combines the power of cloud and edge computing in a single package and enhances the manufacturing output for most manufacturers. It is truly the best of both worlds as it delivers a centralized IT management with strong AI insights on the edge platforms. The journey starts with IoT, AI, and big data, which may be difficult to scale. But by adding edge and cloud computing together onto a single platform, a revolutionary angle can be brought to manufacturing, bringing factory owners closer to Industry 4.0.