Overhaul of the chemical industry
With the growing customer demands, the chemical industry is finding itself drowning in multidimensional challenges. The complexities may arise due to complicated supply chains, uneven demand patterns, regulation and compliance pressures, and even from the lack of digital transformation. The pressing challenge that seems to put a strain on the shop floor operations and operators comes across to be low operational performance which results from low asset reliability and utilization. This issue also stems out from low visibility of the sop floor and operations being conducted in-silo. This leads to increased wastage and costs from the low-quality operations and output. The operators are involved in reworking on the low-quality output or even scrapping them entirely, which is a of result process variations. On top of it, the low degree of digital maturity and lack of smart manufacturing at the plant hinders continuous improvement. The main reason is owed to an insufficient and improper use of process, equipment and data. Further, the untrained operators also add to the unwanted cost of inefficient production processes as they maintain paper based log sheets which complicate the manual calculations and make verification of production information time intensive. Until now, the digital manufacturing operations needed to meet these industry challenges posed as a challenge in itself. Multiple layers of technology are required to connect store floors and IT systems to create meaningful information, deliver quickly, and generate the desired business impact. Solulever, a Dutch technology Startup, developed a manufacturing intelligence and connectivity software platform, called Brabo to fulfil these needs. Brabo is a wholesome solution that is quickly being adopted by major chemical companies. This Industry 4.0 platform works as an integrated manufacturing intelligence system that clients benefit from in areas of resource efficiency, asset utilization, process control and improvement, and OT service management. Brabo provides real-time data transfer capabilities for the shop floor operators to take preventive and a corrective decisions and measures during production processes. Additionally, the Edge computing and IIoT platform proffers customers a drag-and-drop framework to create, configure, and visualize enhanced SCADA systems that replace the need for SCADA system from a third party. Chemical industry leaders are now moving to adopt Brabo, an innovative OT-IT platform for smart manufacturing that makes the digital transformation of manufacturing operations seamless.